requirements.txt
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-telegram-bot==21.1.1
apscheduler==3.10.4
holidays==0.75
